+++ anyblok_pyramid_rest_api/tests/test_crud_resource.py
@@ -45,15 +45,15 @@ class TestCrudResourceBase(PyramidDBTestCase):
 
     def test_example_collection_post(self):
         """Example POST /examples/"""
-        response = self.webserver.post_json('/examples', {'name': 'plip'})
+        response = self.webserver.post_json('/examples', [{'name': 'plip'}])
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.json_body.get('name'), "plip")
+        self.assertEqual(response.json_body[0]['name'], "plip")
 
     def test_example_collection_post_with_errors(self):
         """Example POST /examples/with/errors"""
         with LogCapture() as logs:
             fail = self.webserver.post_json(
-                '/examples/with/errors', {'name': 'plip'}, status=500)
+                '/examples/with/errors', [{'name': 'plip'}], status=500)
 
         self.assertEqual(fail.status_code, 500)
         self.assertEqual(fail.json_body.get('status'), 'error')
@@ -347,34 +347,34 @@ class CrudResourceSchema:
     def test_customer_collection_post(self):
         """Customer POST /customers"""
         response = self.webserver.post_json(self.collection_path,
-                                            {'name': 'plip'})
+                                            [{'name': 'plip'}])
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.json_body.get('name'), "plip")
+        self.assertEqual(response.json_body[0]['name'], "plip")
 
     def test_customer_collection_post_partial(self):
         """Customer POST partial /customers"""
         response = self.webserver.post_json(self.collection_path,
-                                            {'name': 'plip'})
+                                            [{'name': 'plip'}])
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.json_body.get('name'), "plip")
+        self.assertEqual(response.json_body[0]['name'], "plip")
 
     def test_customer_collection_post_empty_body(self):
         """Customer POST empty body /customers"""
         fail = self.webserver.post_json(
-            self.collection_path, {}, status=400)
+            self.collection_path, [{}], status=400)
         self.assertEqual(fail.status_code, 400)
         self.assertEqual(fail.json_body.get('status'), 'error')
         self.assertEqual(
             fail.json_body.get('errors')[0].get('location'), 'body')
         self.assertDictEqual(
             fail.json_body.get('errors')[0].get('description'),
-            {'name': ['Missing data for required field.']}
+            {'0': {'name': ['Missing data for required field.']}}
         )
 
     def test_customer_collection_post_bad_key_in_body(self):
         """Customer POST bad key in body /customers"""
         fail = self.webserver.post_json(
-            self.collection_path, {'unexistingkey': 'plip'}, status=400)
+            self.collection_path, [{'unexistingkey': 'plip'}], status=400)
         self.assertEqual(fail.status_code, 400)
         self.assertEqual(fail.json_body.get('status'), 'error')
         self.assertEqual(
@@ -382,7 +382,19 @@ class CrudResourceSchema:
         self.assertEqual(
             fail.json_body.get('errors')[0].get('name'),
             'Validation error for body')
-        self.assertEqual(len(fail.json_body.get('errors')), 2)
+        self.assertDictEqual(
+            fail.json_body['errors'][0]['description'],
+            {
+                '0': {
+                    'name': ['Missing data for required field.'],
+                    'unexistingkey': [
+                        'Unknown field.',
+                        "Unknown fields {'unexistingkey'} on Model "
+                        "Model.Customer"
+                    ]
+                }
+            }
+        )
 
     def test_customer_put(self):
         """Customer PUT /customers/{id}"""