+++ CHANGELOG.rst
@@ -44,6 +44,8 @@ Refactored
 
 * If a ``request.error`` is found during the execution of a view in the crud resource,
   then a registry.rollback will be done
+* ``Collection post`` is waiting a list of dict, Now it is possible to create more than
+  one entry
 
 0.3.0 (2018-12-07)
 ------------------